jirigo_engine/requirements.txt
autopep8==1.5.1
click==7.1.1
flash==1.0.3
Flask==1.1.2
Flask-Cors==3.0.8
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
numpy==1.18.2
pandas==1.0.3
psycopg2-binary==2.8.5
pycodestyle==2.5.0
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
Werkzeug==1.0.1
